requirements.txt
"""
FastAPI==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
streamlit==1.29.0
yfinance==0.2.31
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
xgboost==2.0.2
prophet==1.1.5
stable-baselines3[extra]==2.3.2
backtrader==1.9.78.123
vectorbt==0.26.2
celery==5.3.6
redis==5.0.1
flower==2.0.1
python-telegram-bot==20.7
schedule==1.2.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
httpx==0.25.2
pytest==7.4.3
black==23.12.1
isort==5.13.2
pre-commit==3.6.0
docker==7.0.0
"""

